move paragraph components shared utils to draft/shared
